app and Dockerfile are modified
